[TS-176] Refactor: directory, file name rechange
