+++ src/main.tsx
@@ -3,8 +3,8 @@ import ReactDOM from "react-dom/client";
 
 import { Global } from "@emotion/react";
 
-import { GlobalStyle } from "@/abc/abc";
 import App from "@/App";
+import { GlobalStyle } from "@/styles/globalStyle";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
 	<React.StrictMode>
